requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.2.5
autopep8==1.4.4
bleach==3.1.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
distlib==0.3.0
Django==3.0.7
django-appconf==1.0.4
django-avatar==5.0.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-imperavi==0.2.3
django-markdownx==3.0.1
-e git://github.com/dokterbob/django-newsletter.git@4a518f5380ed00602fa83b14b1d52842f0845a7a#egg=django_newsletter
django-phone-field==1.8.1
django-social-share==2.1.1
django-taggit==1.2.0
django-widget-tweaks==1.4.8
docutils==0.16
filelock==3.0.12
gunicorn==20.1.0
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
keyring==21.2.0
lazy-object-proxy==1.4.1
ldif3==3.1.1
Markdown==3.3.3
mccabe==0.6.1
pdfpackage==1.0
Pillow==8.0.1
pipenv==2018.11.26
pkginfo==1.5.0.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.3.1
PyMySQL==1.0.2
python-card-me==0.9.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
rope==0.16.0
six==1.12.0
sorl-thumbnail==12.7.0
sqlparse==0.3.1
surlex==0.2.0
tqdm==4.44.1
twine==3.1.1
typed-ast==1.4.0
unicodecsv==0.14.1
urllib3==1.25.8
virtualenv==20.0.14
virtualenv-clone==0.5.3
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.11.2
zipp==3.1.0
